refactor: Move LogoutCommand and LogoutCommandHandler to application layer and update imports
