--- frontend/src/modules/auth/app/LogoutCommand.ts
@@ -1,8 +1,0 @@
-/**
- * Command to represent logout user intent in auth context.
- * Following CQRS pattern for authentication operations.
- */
-export class LogoutCommand {
-  // No parameters needed for logout
-  constructor() {}
-}
