+++ frontend/src/modules/auth/application/commands/LogoutCommand.ts
@@ -1,0 +1,4 @@
+export class LogoutCommand {
+  // No parameters needed for logout
+  constructor() {}
+}
